+++ src/backend/utils/adt/tsvector_op.c
@@ -1149,7 +1149,7 @@ static TSVectorStat *
 ts_stat_sql(MemoryContext persistentContext, text *txt, text *ws)
 {
 	char	   *query = text_to_cstring(txt);
-	int64			i;
+	int			i;
 	TSVectorStat *stat;
 	bool		isnull;
 	Portal		portal;
